requirements.txt
aiomysql
discord.py